+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 22/11/2025
 
 - Added No Loops rule
+- You can now filter collections by type
 - Fixed solvers not aborting properly
 - Improved comment sidebar UI